project-root/backend/requirements.txt
fastapi
uvicorn[standard]
websockets
sqlalchemy
asyncpg
passlib
bcrypt
python-jose
